blbl_md5/requirements.txt
mysql-connector-python~=9.3.0
requests~=2.32.3